@@ -17,7 +17,7 @@
 package org.tensorflow.lite.examples.posenet
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 
 class CameraActivity : AppCompatActivity() {
 
